clion adjustment

+++ include/NodeData.h
@@ -9,8 +9,8 @@
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
 
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -95,7 +95,7 @@ struct NodeData {
     size_t id_art_var_convex;
     int    id_art_var_assignment;
     size_t id_art_var_cuts;
-    size_t id_next_var_cuts;
+    int    id_next_var_cuts;
     int    id_pseudo_schedules;
 
     // PricerSolver
